Move test configuration eg. API_BASE_URL and ENDPOINT_PATH to configuration.py files.
