+++ api/tests/python/end_points/auto_analyse/configuration.py
@@ -1,0 +1,2 @@
+API_BASE_URI = '/api/v1/'
+ENDPOINT_PATH = API_BASE_URI + 'name-analysis'
