+++ api/tests/python/end_points/auto_analyse/protected_names/test_add_descriptive_word_both_classifications_request.py
@@ -7,7 +7,7 @@ from namex.services.name_request.auto_analyse import AnalysisIssueCodes
 
 from ..common import assert_issues_count_is_gt, assert_issue_type_is_one_of, assert_has_word_upper, \
     save_words_list_classification
-from ..common import ENDPOINT_PATH
+from ..configuration import ENDPOINT_PATH
 from ...common import token_header, claims
