+++ api/tests/python/end_points/auto_analyse/protected_names/test_corporate_name_conflict_strip_out_numbers_request.py
@@ -7,7 +7,7 @@ from namex.services.name_request.auto_analyse import AnalysisIssueCodes
 
 from ..common import assert_issues_count_is_gt, assert_correct_conflict, save_words_list_name, \
     save_words_list_classification, assert_additional_conflict_parameters
-from ..common import ENDPOINT_PATH
+from ..configuration import ENDPOINT_PATH
 from ...common import token_header, claims
