+++ api/tests/python/end_points/auto_analyse/protected_names/test_name_use_special_words_more_than_one_request.py
@@ -7,7 +7,7 @@ from namex.services.name_request.auto_analyse import AnalysisIssueCodes
 
 from ..common import assert_issues_count_is_gt, assert_has_word_upper, save_words_list_classification, \
     save_words_list_virtual_word_condition
-from ..common import ENDPOINT_PATH
+from ..configuration import ENDPOINT_PATH
 from ...common import token_header, claims
