+++ api/tests/python/end_points/auto_analyse/protected_names/test_request_actions_bc.py
@@ -4,7 +4,7 @@ import jsonpickle
 from urllib.parse import quote_plus
 
 from ..common import save_words_list_classification, assert_issues_count_is_zero
-from ..common import ENDPOINT_PATH
+from ..configuration import ENDPOINT_PATH
 from ...common import token_header, claims
