+++ api/tests/python/end_points/auto_analyse/protected_names/test_successful_well_formed_request.py
@@ -4,7 +4,7 @@ import jsonpickle
 from urllib.parse import quote_plus
 
 from ..common import save_words_list_classification
-from ..common import ENDPOINT_PATH
+from ..configuration import ENDPOINT_PATH
 from ...common import token_header, claims
